refactor: mudanca de biblioteca de mapeamento de Dtos
