--- microservice02/src/main/java/com/silmarils/microservice02/dto/mapper/PostMapper.java
@@ -1,19 +1,0 @@
-package com.silmarils.microservice02.dto.mapper;
-
-import com.silmarils.microservice02.dto.PostCreateDto;
-import com.silmarils.microservice02.dto.PostResponseDto;
-import com.silmarils.microservice02.entities.Post;
-import org.mapstruct.Mapper;
-
-import java.util.List;
-
-@Mapper(componentModel = "spring")
-public interface PostMapper {
-
-    Post postDtoToPost (PostCreateDto postCreateDto);
-
-    PostResponseDto postToPostResponseDto (Post post);
-
-    List<PostResponseDto> postToPostResponseDtoList(List<Post> posts);
-
-}
