+++ microservice02/src/main/java/com/silmarils/microservice02/web/dto/PostResponseDto.java
@@ -1,4 +1,4 @@
-package com.silmarils.microservice02.dto;
+package com.silmarils.microservice02.web.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,6 +10,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PostResponseDto {
+    private String id;
     private Integer userId;
     private String title;
     private String body;
